requirements.txt
Flask
requests
deepface
pillow